fix gson, fastkv add null check

+++ enoa-patch/enoa-json-gson/src/main/java/io/enoa/json/provider/gson/GNNumber.java
@@ -111,7 +111,8 @@ class GNNumber {
         return false;
     }
     try {
-      Double.parseDouble(text);
+//      Double.parseDouble(text);
+      Long.parseLong(text);
     } catch (Exception e) {
       return false;
     }
